tests: Add MarshalByRefObject
This commit adds tests for MarshalByRefObjects, it is a direct copy of
the ExportInterfaceTest, modified slightly.

It immediately shows some divergence in behaviour between the two
methods of object export.

This class requires a tidy to avoid name clashes, we solve that in the
immediate future by placing these tests into a new namespace.