tests: ObjectManager like interface
This commit adds a short test for an object exporting an object manager
like interface. This commit may introduce some repetition of the
ExportInterfaceTest but that is quite large already. We will use this
new test to isolate and eventually fix an issue when connecting to
signals on such an interface.