build: Add Makefile.include.in
This commit adds support for release configuration through a configure
switch, --enable-release. For backwards compatibility this is left off
by default.

We also update the test run target with the path to the built assembly,
based on the previous commit placing each project's output in a
directory according to convention.